Trop de changements

+++ traitement/upload_com.php
@@ -1,16 +1,31 @@
 <?php
+        $increment = 0;
         $target_dir = "uploads/coms/";
         $target_file = $target_dir . basename($_FILES["imageCom"]["name"]);
         $uploadOk = 1;
-        $imageComFileType = strtolower(pathinfo($target_file,PATHINFO_EXTENSION));
+        $imageFileType = strtolower(pathinfo($target_file,PATHINFO_EXTENSION));
         if ($uploadOk == 0) {
-            echo "Désolé, votre imageCom n'a pas été publié.";
+            echo "Désolé, votre image n'a pas été publié.";
        
         } else {
-            if (move_uploaded_file($_FILES["imageCom"]["tmp_name"], $target_file)) {
-                echo "Votre imageCom". htmlspecialchars( basename( $_FILES["imageCom"]["name"])). " a bien été publié.";
-            } else {
-                echo "Désolé, votre imageCom n'a pas été publié.";
+            if(file_exists($target_file)) {
+                while(file_exists($target_file)) {
+                $increment++;
+                $target_file = $target_dir.$increment.basename($_FILES["imageCom"]["name"]);
             }
+            
+                if (move_uploaded_file($_FILES["imageCom"]["tmp_name"], $target_file)) {
+                    echo "Votre image". htmlspecialchars( basename( $_FILES["imageCom"]["name"])). " a bien été publié.";
+                } else {
+                    echo "Désolé, votre image n'a pas été publié.";
+                }
+            }else{
+                if (move_uploaded_file($_FILES["imageCom"]["tmp_name"], $target_file)) {
+                    echo "Votre image". htmlspecialchars( basename( $_FILES["imageCom"]["name"])). " a bien été publié.";
+                } else {
+                    echo "Désolé, votre image n'a pas été publié.";
+                }
+            }
+            
         }
 ?>
