+++ traitement/upload_post.php
@@ -1,4 +1,5 @@
 <?php
+        $increment = 0;
         $target_dir = "uploads/post/";
         $target_file = $target_dir . basename($_FILES["image"]["name"]);
         $uploadOk = 1;
@@ -7,10 +8,24 @@
             echo "Désolé, votre image n'a pas été publié.";
        
         } else {
-            if (move_uploaded_file($_FILES["image"]["tmp_name"], $target_file)) {
-                echo "Votre image". htmlspecialchars( basename( $_FILES["image"]["name"])). " a bien été publié.";
-            } else {
-                echo "Désolé, votre image n'a pas été publié.";
+            if(file_exists($target_file)) {
+                while(file_exists($target_file)) {
+                $increment++;
+                $target_file = $target_dir.$increment.basename($_FILES["image"]["name"]);
             }
+            
+                if (move_uploaded_file($_FILES["image"]["tmp_name"], $target_file)) {
+                    echo "Votre image". htmlspecialchars( basename( $_FILES["image"]["name"])). " a bien été publié.";
+                } else {
+                    echo "Désolé, votre image n'a pas été publié.";
+                }
+            }else{
+                if (move_uploaded_file($_FILES["image"]["tmp_name"], $target_file)) {
+                    echo "Votre image". htmlspecialchars( basename( $_FILES["image"]["name"])). " a bien été publié.";
+                } else {
+                    echo "Désolé, votre image n'a pas été publié.";
+                }
+            }
+            
         }
 ?>
